+++ mango3-leptos-utils/src/models/mod.rs
@@ -1,7 +1,9 @@
 mod action_form_resp;
 mod basic_config_resp;
+mod blob_resp;
 mod user_resp;
 
 pub use action_form_resp::ActionFormResp;
 pub use basic_config_resp::BasicConfigResp;
+pub use blob_resp::BlobResp;
 pub use user_resp::UserResp;