+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Basic setup.
 - Feature to create users.
 - Features to login and logout.
+- Features to upload and show images.